+++ R/crawford.test.R
@@ -10,7 +10,10 @@
 #' @return Returns a data frame containing the t-value, degrees of freedom, and p-value. If significant, the case is different from the control group.
 #'
 #' @examples
+#' library(psycho)
+#'
 #' crawford.test(case = 10, controls = c(0, -2, 5, 2, 1, 3, -4, -2))
+#' crawford.test(case = 7, controls = c(0, -2, 5, 2, 1, 3, -4, -2))
 #'
 #' @author Dan Mirman, Dominique Makowski
 #'
@@ -23,21 +26,38 @@ crawford.test <- function(case, controls, verbose=T) {
 
   pval <- 2 * (1 - pt(abs(tval), df = degfree)) # two-tailed p-value
 
+  # One-tailed p value
+  if(pval > .05 & pval/2 < .05){
+    one_tailed <- paste0(" However, the null hypothesis of no difference can be rejected at a one-tailed 5% significance level (one-tailed p ",
+                         format_p(pval/2),
+                         ").")
+
+  } else {
+    one_tailed <- ""
+  }
+
+
   p_interpretation <- ifelse(pval < 0.05, " significantly ", " not significantly ")
-  t_interpretation <- ifelse(tval < 0, "below ", "above ")
-  pop_interpretation <- ifelse(tval < 0, " above ", " below ")
+  t_interpretation <- ifelse(tval < 0, " lower than ", " higher than ")
 
   text <- paste0(
-    "The Crawford-Howell (1998) t-test suggests that the patient is",
+    "The Crawford-Howell (1998) t-test suggests that the patient's score (",
+    format_digit(case),
+    ") is",
     p_interpretation,
-    t_interpretation,
-    "the control group (t(",
+    "different from the controls (M = ",
+    format_digit(mean(controls)),
+    ", SD = ",
+    format_digit(sd(controls)),
+    ", t(",
     degfree,
     ") = ",
     format_digit(tval),
     ", p ",
     format_p(pval),
-    "). It has estimated that the patient is located",
+    ").",
+    one_tailed,
+    " The patient's score is",
     t_interpretation,
     format_digit((1 - pval) * 100),
     "% of the control population."