+++ R/mellenbergh.test.R
@@ -10,6 +10,8 @@
 #' @return Returns a data frame containing the z-value and p-value. If significant, the difference between pre and post tests is significant.
 #'
 #' @examples
+#' library(psycho)
+#'
 #' mellenbergh.test(t0 = 4, t1 = 12, controls = c(0, -2, 5, 2, 1, 3, -4, -2))
 #' mellenbergh.test(t0 = 8, t1 = 2, controls = 2.6)
 #'