+++ NEWS.md
@@ -3,12 +3,16 @@
 ### Breaking changes
 - Changed order of parameters in `dprime`
 ### New functions / parameters
+- Added `find_combinations` function
+- Added `find_best_combination` function
 - Added `is.standardized` function
 - Added `get_contrasts.stanreg` and `get_predicted.stanreg` functions
 - Added `crawford_dissociation.test` function for single-case tests
 - Externalized the "mpe" function
 - Added `affective` dataset
 ### Changes
+- Added random effects summary in `analyze.stanreg`
+- Improved `mellenbergh.test`
 - Improved `analyze.stanreg`
 - Added new vignette for "Bayesian analyses in psychology"
 - Improved `crawford.test`